Consolidation of rmGame and rmGame2 assets
rmGame was always supposed to be the only room where game levels
happened.  rmGame2 was an experimental room for 2-planet levels.  Using
it introduced a lot of redundant code and complexity that the project
could do without.  Eliminated rmGame2 and re-wrote the level reset code
to handle incrementing the number of planets.