Removed bullet self-destruct
Realized that if you can self-destruct your bullet, you lose the risk of
accidentally shooting yourself, which is literally the only danger in
the game.  Self-destruct only works in debug_mode for now; if I
re-introduce this feature it will only be once there are risks to the
player's life.