Improved planet cluster balancing
Fixed some bugs affecting the planet clutser re-centering with multiple
planets.  When planets are spawned that are touching each other, they
will re-position themselves, clamp themselves to within the room such
that they are at least 100px from a window border, and the entire
cluster will shift their offset so that their average position is the
center of the room.  This mostly works, but sometimes I am still seeing
them not be perfectly centered, and I haven't been able to figure out
why, but this behavior is still much better than it has been.