Remove timer, time bonus
Removed level timer, time bonus from game. Time is not in the user's
control due to varying speed of rotation. Game design intent is to
reward patience and careful aim, so a countdown timer is antithetical to
the design.

This will be replaced with a par bonus system based on the number of
shots fired.